Made the form Component reusable except for partners
