--- src/pages/AdminPage/components/FormComponent.jsx
@@ -1,41 +1,0 @@
-import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { sendFormData } from './redux/actions';
-
-const FormComponent = () => {
-  const [title, setTitle] = useState('');
-  const [description, setDescription] = useState('');
-  const [file, setFile] = useState(null);
-  const dispatch = useDispatch();
-  const token = 'your_token_here';
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    const formData = new FormData();
-    formData.append('title', title);
-    formData.append('description', description);
-    formData.append('file', file);
-
-    dispatch(sendFormData(formData, token));
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <div>
-        <label>Title:</label>
-        <input type="text" value={title} onChange={(e) => setTitle(e.target.value)} />
-      </div>
-      <div>
-        <label>Description:</label>
-        <textarea value={description} onChange={(e) => setDescription(e.target.value)} />
-      </div>
-      <div>
-        <label>File:</label>
-        <input type="file" onChange={(e) => setFile(e.target.files[0])} />
-      </div>
-      <button type="submit">Submit</button>
-    </form>
-  );
-};
-
-export default FormComponent;
